Updated Lab-text-skills.md
Added content to this lab to hand hold students in the use of POSTMAN in step 1. Feedback from the India bootcamp had students struggling with this. Therefore, the objective is to reduce the amount of student support required in POSTMAN when they do this lab, as they have detailed instructions